Fixed negative values to be displayed for Hysterie and Auslegungstemp

+++ src/ems.h
@@ -269,7 +269,7 @@ typedef struct {
     uint8_t auslegungstemp ; // temperature bei min temp
     uint8_t maxvorlauf ; // maximale Vorlauf Temp
     uint8_t roomoffset ; // temperature bei min temp
-    uint8_t minoutsidetemp; // minimal temp in the region
+    int minoutsidetemp; // minimal temp in the region
     uint8_t housetype; // light medium heavy house types
     bool tempaveragebool; // outside temp is mixed to average temp
     // lobocobra end
